+++ admin/src/models/Plugin.ts
@@ -1,8 +1,0 @@
-/**
- * Defines the structure of a plugin object.
- */
-export interface Plugin {
-  readonly id: string
-  readonly name: string
-  readonly description: string
-}