+++ core/cat/looking_glass/agent_manager.py
@@ -1,23 +1,20 @@
 # import re
 from copy import copy
 
-from cat.utils import log
+from cat.log import log
 from langchain.agents import AgentExecutor, ConversationalAgent
 from langchain.chains import LLMChain
 
 
 class AgentManager:
     def __init__(self, cat):
-        self.verbose = cat.verbose
         self.cat = cat
 
     def get_agent_executor(self):
         mad_hatter = self.cat.mad_hatter
 
         prompt_prefix = mad_hatter.execute_hook("agent_prompt_prefix")
-        prompt_format_instructions = mad_hatter.execute_hook(
-            "agent_prompt_instructions"
-        )
+        prompt_format_instructions = mad_hatter.execute_hook("agent_prompt_instructions")
         prompt_suffix = mad_hatter.execute_hook("agent_prompt_suffix")
 
         # extract automatically input_variables from prompt parts (can't do it yet)
@@ -34,9 +31,7 @@ class AgentManager:
             "agent_scratchpad",
         ]
 
-        allowed_tools = mad_hatter.execute_hook(
-            "agent_allowed_tools", copy(mad_hatter.tools)
-        )
+        allowed_tools = mad_hatter.execute_hook("agent_allowed_tools", copy(mad_hatter.tools))
         allowed_tools_names = [t.name for t in allowed_tools]
 
         prompt = ConversationalAgent.create_prompt(
@@ -49,26 +44,21 @@ class AgentManager:
             input_variables=input_variables,
         )
 
-        if self.verbose:
-            log("Using prompt template:")
-            log(prompt.template)
+        log("Sending prompt", "INFO")
+        log(prompt.template, "DEBUG")
 
         # main chain
-        agent_chain = LLMChain(prompt=prompt, llm=self.cat.llm, verbose=self.verbose)
+        agent_chain = LLMChain(prompt=prompt, llm=self.cat.llm, verbose=True)
 
         # init agent
         agent = ConversationalAgent(
             llm_chain=agent_chain,
             allowed_tools=allowed_tools_names,
-            verbose=self.verbose,
         )
 
         # agent executor
         agent_executor = AgentExecutor.from_agent_and_tools(
-            agent=agent,
-            tools=allowed_tools,
-            return_intermediate_steps=True,
-            verbose=self.verbose,
+            agent=agent, tools=allowed_tools, return_intermediate_steps=True
         )
 
         return agent_executor