+++ core/cat/mad_hatter/core_plugin/hooks/models.py
@@ -1,8 +1,15 @@
+"""Hooks to modify the Cat's language and embedding models.
+
+Here is a collection of methods to hook into the settings of the Large Language Model and the Embedder.
+
+"""
+
 import os
 
 import cat.factory.llm as llms
 import cat.factory.embedder as embedders
 from cat.db import crud
+from langchain.llms.base import BaseLLM
 from langchain.llms import Cohere, OpenAI, OpenAIChat, AzureOpenAI
 from langchain import HuggingFaceHub
 from langchain.chat_models import AzureChatOpenAI
@@ -10,7 +17,21 @@ from cat.mad_hatter.decorators import hook
 
 
 @hook(priority=0)
-def get_language_model(cat):
+def get_language_model(cat) -> BaseLLM:
+    """Hook into the Large Language Model (LLM) selection.
+
+    Allows to modify how the Cat selects the LLM at bootstrap time.
+
+    Bootstrapping is the process of loading the plugins, the natural language objects (e.g. the LLM),
+    the memories, the *Agent Manager* and the *Rabbit Hole*.
+
+    Args:
+        cat: Cheshire Cat instance.
+
+    Returns:
+        langchain `BaseLLM` instance for the selected model.
+
+    """
     selected_llm = crud.get_setting_by_name(next(cat.db()), name="llm_selected")
 
     if selected_llm is None:
@@ -33,6 +54,19 @@ def get_language_model(cat):
 
 @hook(priority=0)
 def get_language_embedder(cat):
+    """Hook into the  embedder selection.
+
+    Allows to modify how the Cat selects the embedder at bootstrap time.
+
+    Bootstrapping is the process of loading the plugins, the natural language objects (e.g. the LLM),
+    the memories, the *Agent Manager* and the *Rabbit Hole*.
+
+    Args:
+        cat: Cheshire Cat instance.
+
+    Returns:
+        Selected embedder model.
+    """
     # Embedding LLM
 
     print("naked cat: ", cat.llm)