+++ core/cat/main.py
@@ -1,8 +1,9 @@
-from cat.utils import logger
+from cat.log import log
 import os
 from contextlib import asynccontextmanager
 
 from fastapi import Depends, FastAPI
+from fastapi.staticfiles import StaticFiles
 from cat.routes import base, memory, plugins, upload, websocket
 from cat.api_auth import check_api_key
 from fastapi.responses import JSONResponse
@@ -16,6 +17,7 @@ from cat.looking_glass.cheshire_cat import CheshireCat
 # logging.debug('EXAMPLE with debug category')
 # logging.info('EXAMPLE with info category')
 # logging.error('EXAMPLE with error category')
+
 
 @asynccontextmanager
 async def lifespan(app: FastAPI):
@@ -47,22 +49,19 @@ cheshire_cat_api.add_middleware(
 
 # Add routers to the middleware stack.
 cheshire_cat_api.include_router(base.router, tags=["Status"])
-cheshire_cat_api.include_router(
-    general_setting.router, tags=["Settings - General"], prefix="/settings"
-)
-cheshire_cat_api.include_router(
-    llm_setting.router, tags=["Settings - Large Language Model"], prefix="/settings/llm"
-)
-cheshire_cat_api.include_router(
-    embedder_setting.router, tags=["Settings - Embedder"], prefix="/settings/embedder"
-)
+cheshire_cat_api.include_router(general_setting.router, tags=["Settings - General"], prefix="/settings")
+cheshire_cat_api.include_router(llm_setting.router, tags=["Settings - Large Language Model"], prefix="/settings/llm")
+cheshire_cat_api.include_router(embedder_setting.router, tags=["Settings - Embedder"], prefix="/settings/embedder")
 cheshire_cat_api.include_router(plugins.router, tags=["Plugins"], prefix="/plugins")
 cheshire_cat_api.include_router(memory.router, tags=["Memory"], prefix="/memory")
-cheshire_cat_api.include_router(
-    upload.router, tags=["Rabbit Hole (file upload)"], prefix="/rabbithole"
-)
+cheshire_cat_api.include_router(upload.router, tags=["Rabbit Hole (file upload)"], prefix="/rabbithole")
 cheshire_cat_api.include_router(websocket.router, tags=["Websocket"])
 
+# admin (static build)
+cheshire_cat_api.mount("/admin", StaticFiles(directory="/admin/dist/", html=True), name="admin")
+
+# static files
+# TODO
 
 # error handling
 @cheshire_cat_api.exception_handler(RequestValidationError)